requirements.txt
Django==3.2.*
djangorestframework==3.12.*
djangorestframework-jwt==1.11.*
psycopg2-binary==2.9.*
